backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pandas
python-dotenv
xgboost
scikit-learn
